+++ frontend/src/hooks/useLiveTickerFeed.ts
@@ -1,11 +1,10 @@
-import { useAlpacaDataProvider } from "@/services/MarketDataProvider/MarketDataContext";
+import api from "@/api/client";
 import { OHLCVData, Ticker, LiveTickerFeedMessage } from "@shared/types";
 import { useState, useEffect, useCallback } from "react";
 
 export function useLiveTickerFeed(
   tickers: Ticker[],
 ): Record<string, LiveTickerFeedMessage> {
-  const provider = useAlpacaDataProvider();
   const [messages, setMessages] = useState<Record<string, LiveTickerFeedMessage>>({});
 
   const onTick = useCallback((data: OHLCVData) => {
@@ -30,14 +29,13 @@ export function useLiveTickerFeed(
   }, []);
 
   useEffect(() => {
-    if (!provider?.subscribeTickers || !provider?.unsubscribeTickers) return;
     if (tickers.length === 0) return;
 
-    provider.subscribeTickers(tickers, onTick);
+    api.subscribeTickers(tickers, onTick);
     return () => {
-      provider.unsubscribeTickers!(tickers, onTick);
+      api.unsubscribeTickers(tickers, onTick);
     };
-  }, [provider, onTick, JSON.stringify(tickers.map((t) => t.symbol))]);
+  }, [onTick, JSON.stringify(tickers.map((t) => t.symbol))]);
 
   return messages;
 }
