--- frontend/src/services/MarketDataProvider/providers/index.ts
@@ -1,1 +1,0 @@
-export { AlpacaProvider } from "./alpaca.ts";
